Bump yangtools to 4.0.1
This bumps yangtools to 4.0.1 to pick up Uint64 fix.

+++ dom/mdsal-dom-spi/src/test/java/org/opendaylight/mdsal/dom/spi/store/AbstractDOMStoreTreeChangePublisherTest.java
@@ -64,7 +64,7 @@ public class AbstractDOMStoreTreeChangePublisherTest extends AbstractDOMStoreTre
         final Field closedField = AbstractRegistration.class.getDeclaredField("closed");
         closedField.setAccessible(true);
 
-        final int closed = (int) closedField.get(abstractDOMDataTreeChangeListenerRegistration);
+        final byte closed = (byte) closedField.get(abstractDOMDataTreeChangeListenerRegistration);
         Assert.assertEquals(1, closed);
     }
